pyproject.toml
[project]
name = "django_2"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jungminchae", email = "minchae3618@gmail.com"},
]
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15.2",
    "django-ninja>=1.2.2",
    "uvicorn>=0.30.5",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
